* compat/strtod.c (strtod): simplified #includes         * compat/strtol.c (strtol): gather result in a long before returning         as a long: necessary on platforms where sizeof(int) != sizeof(long).

+++ compat/strtol.c
@@ -54,7 +54,7 @@ strtol(string, endPtr, base)
 				 */
 {
     register CONST char *p;
-    int result;
+    long result;
 
     /*
      * Skip any leading blanks.
